--- src/main/java/com/marginallyclever/convenience/log/LogListener.java
@@ -1,6 +1,0 @@
-package com.marginallyclever.convenience.log;
-
-public interface LogListener {
-	// decorated HTML message has been logged
-	public void logEvent(String message);
-}
