--- src/main/java/com/marginallyclever/convenience/log/LogPanelListener.java
@@ -1,5 +1,0 @@
-package com.marginallyclever.convenience.log;
-
-public interface LogPanelListener {
-	public void commandFromLogPanel(String msg);
-}
